auto-claims-demo/ai-service/pyproject.toml
[project]
name = "auto-claims-ai-service"
version = "0.1.0"
description = "Main AI service for the Auto Claims Demo, orchestrating damage detection, estimation, and other agentic tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.4",
    "google-cloud-aiplatform>=1.136.0",
    "google-cloud-vision>=3.12.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "google-genai",
    "google-adk[a2a,bigquery-analytics]==1.29.0",
    "google-cloud-storage>=3.9.0",
    "google-cloud-bigquery",
    "ultralytics>=8.4.13",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.0",
    "huggingface-hub>=1.4.1",
    "opentelemetry-instrumentation-google-genai>=0.1.0",
    "google-cloud-logging>=3.12.0",
    "opentelemetry-instrumentation-httpx>=0.59b0",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "opentelemetry-propagator-gcp>=1.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]
